requirements.txt
Flask
whisper
gtts
pygame
python-dotenv
numpy
sounddevice
scipy
soundfile